requirements.txt
aiodns==4.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
annotated-types==0.7.0
anyio==4.12.1
asyncio-mqtt==0.16.2
attrs==25.4.0
black==26.1.0
build==1.4.0
cachetools==7.0.2
ccxt==4.5.40
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coincurve==21.0.0
colorlog==6.10.1
cryptography==46.0.5
dependency-injector==4.48.3
dingtalk-sdk==1.3.8
feedparser==6.0.12
flake8==7.3.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
librt==0.8.1
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
msgpack==1.1.2
multidict==6.7.1
mypy==1.19.1
mypy_extensions==1.1.0
numpy==2.4.2
optionaldict==0.1.2
packaging==26.0
paho-mqtt==2.1.0
pandas==3.0.1
pathspec==1.0.4
platformdirs==4.9.2
pluggy==1.6.0
propcache==0.4.1
psutil==7.2.2
pycares==5.0.1
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
pyflakes==3.4.0
Pygments==2.19.2
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyproject_hooks==1.2.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-html==4.2.0
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytokens==0.4.1
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.2.1
requests==2.32.5
rumps==0.4.0
schedule==1.2.2
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.0
sgmllib3k==1.0.0
six==1.17.0
SQLAlchemy==2.0.48
TA-Lib==0.6.8
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
